utils/requirements.txt
python-gnupg==0.5.0
google-cloud-kms==2.16.1